refactor(droidcall): keep analyzer self-contained
